pkg/nodeplugins/markdown: restructure nodeplugins/markdown package
