+++ pkg/nodeplugins/markdown/document_worker_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package document_test
+package markdown_test
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 
 	"github.com/gardener/docforge/cmd/hugo"
 	"github.com/gardener/docforge/pkg/manifest"
-	"github.com/gardener/docforge/pkg/nodeplugins/markdown/document"
+	document "github.com/gardener/docforge/pkg/nodeplugins/markdown"
 	"github.com/gardener/docforge/pkg/nodeplugins/markdown/linkresolver"
 	"github.com/gardener/docforge/pkg/registry"
 	"github.com/gardener/docforge/pkg/registry/repositoryhost"
