+++ pkg/nodeplugins/markdown/frontmatter/frontmatter_test.go
@@ -12,8 +12,8 @@ import (
 	_ "embed"
 
 	"github.com/gardener/docforge/pkg/manifest"
-	"github.com/gardener/docforge/pkg/nodeplugins/markdown/document/frontmatter"
-	"github.com/gardener/docforge/pkg/nodeplugins/markdown/document/frontmatter/frontmatterfakes"
+	"github.com/gardener/docforge/pkg/nodeplugins/markdown/frontmatter"
+	"github.com/gardener/docforge/pkg/nodeplugins/markdown/frontmatter/frontmatterfakes"
 	"github.com/gardener/docforge/pkg/registry"
 	"github.com/gardener/docforge/pkg/registry/repositoryhost"
