+++ pkg/nodeplugins/markdown/tests/docs/images/gardener-docforge-logo.png
