+++ pkg/nodeplugins/markdown/tests/docs/manifest.yaml
